code push

+++ app/Events/BidPlaced.php
@@ -37,11 +37,11 @@ class BidPlaced implements ShouldBroadcast
         return [
             'bid_id'       => $this->bid->id,
             'ra_name'      => $this->bid->user->name,
-            'bid_amount'   => $this->bid->bid_amount,
-            'total_npv'    => $this->bid->total_npv,
+            'bid_amount'   => (float) $this->bid->bid_amount,
+            'total_npv'    => (float) $this->bid->total_npv,
             'placed_at'    => $this->bid->created_at->format('d M Y, h:i A'),
-            'highest_bid'  => $highestBid,
-            'minimum_next' => $highestBid + $incrementAmount,
+            'highest_bid'  => (float) $highestBid,
+            'minimum_next' => (float) ($highestBid + $incrementAmount),
             'total_bids'   => $auction->bids()->count(),
             'valid_bids'   => $auction->bids()->where('status', 'confirmed')->count(),
         ];
